fix issues Relates #16

+++ server/database/queries/deleteProductQuery.js
@@ -1,5 +1,5 @@
 const connection = require('../connection');
 
-const deleteProductQuery = (productId) => connection.query('DELETE FROM cart WHERE product_id = $1', [productId]);
+const deleteProductQuery = (productId, userId) => connection.query('DELETE FROM cart WHERE product_id = $1 AND user_id = $2', [productId, userId]);
 
 module.exports = deleteProductQuery;
